azure: Refactoring
