+++ serverless/azure/package.json
@@ -4,8 +4,8 @@
   "devDependencies": {
     "@types/node": "^18.x",
     "esbuild": "^0.15.11",
-    "typescript": "^4.8.4",
-    "rimraf": "^5.0.0"
+    "rimraf": "^5.0.0",
+    "typescript": "^4.8.4"
   },
   "private": true,
   "scripts": {
@@ -16,8 +16,10 @@
     "start": "func start"
   },
   "dependencies": {
-    "@azure/functions": "^4.0.0-alpha.7",
-    "fflate": "^0.8.0",
+    "@azure/functions": "^4.1.0",
+    "@azure/identity": "^4.0.0",
+    "@azure/storage-blob": "^12.17.0",
+    "pmtiles": "file:../../js",
     "zlib": "^1.0.5"
   },
   "main": "dist/serverless/azure/src/index.js"
